Update bridge/service-control sample to NServiceBus 10
