+++ samples/bridge/service-control/Bridge_4/Endpoint/Program.cs
@@ -1,14 +1,14 @@
 using System;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
 
+Console.Title = "Endpoint";
+
 var host = Host.CreateDefaultBuilder(args)
-    .ConfigureServices((hostContext, services) =>
-    {
-        Console.Title = "Endpoint";
-        services.AddHostedService<InputLoopService>();
-    })
     .UseNServiceBus(x =>
     {
         var endpointConfiguration = new EndpointConfiguration(
@@ -39,4 +39,37 @@ var host = Host.CreateDefaultBuilder(args)
     })
     .Build();
 
-await host.RunAsync();
+await host.StartAsync();
+
+var messageSession = host.Services.GetRequiredService<IMessageSession>();
+
+const string letters = "ABCDEFGHIJKLMNOPQRSTUVXYZ";
+var random = new Random();
+Console.WriteLine("Press enter to exit");
+Console.WriteLine("Press 'o' to send a message");
+Console.WriteLine("Press 'f' to toggle simulating of message processing failure");
+
+while (true)
+{
+    var key = Console.ReadKey();
+    Console.WriteLine();
+    if (key.Key == ConsoleKey.Enter)
+    {
+        break;
+    }
+    var lowerInvariant = char.ToLowerInvariant(key.KeyChar);
+    switch (lowerInvariant)
+    {
+        case 'o':
+            var id = string.Concat(Enumerable.Range(0, 4).Select(x => letters[random.Next(letters.Length)]));
+            var message = new MyMessage(id);
+            await messageSession.Send(message);
+            break;
+        case 'f':
+            FailureSimulator.Enabled = !FailureSimulator.Enabled;
+            Console.WriteLine($"Failure simulation is now turned {(FailureSimulator.Enabled ? "on" : "off")}");
+            break;
+    }
+}
+
+await host.StopAsync();
